requirements.txt
chromadb
google-generativeai
streamlit
scikit-learn
pandas
numpy
plotly